+++ src/components/admin-monaco.tsx
@@ -1,12 +1,7 @@
 "use client";
 
-import { loader } from "@monaco-editor/react";
-import * as monaco from "monaco-editor";
+import "@/lib/monaco-setup";
 import Editor, { type EditorProps } from "@monaco-editor/react";
-import { configureMonacoEnvironment } from "@/lib/monaco-environment";
-
-configureMonacoEnvironment();
-loader.config({ monaco });
 
 export type AdminMonacoEditorProps = EditorProps;